+++ firebase/extensions/easy-extension/functions/src/utils.ts
@@ -1,6 +1,6 @@
-import { DocumentReference, DocumentSnapshot, FieldValue } from "firebase-admin/firestore";
+import { DocumentReference, DocumentSnapshot, FieldValue, WriteResult } from "firebase-admin/firestore";
 import * as functions from "firebase-functions";
-import config from "./config";
+import { Config } from "./config";
 
 export enum ChangeType {
   CREATE,
@@ -9,7 +9,7 @@ export enum ChangeType {
 }
 
 
-export const getChangeType = (change: functions.Change<DocumentSnapshot>) => {
+export const getChangeType = (change: functions.Change<DocumentSnapshot>): ChangeType => {
   if (!change.after.exists) {
     return ChangeType.DELETE;
   }
@@ -20,9 +20,9 @@ export const getChangeType = (change: functions.Change<DocumentSnapshot>) => {
 };
 
 
-export const success = async (ref: DocumentReference, response: Record<string, any>) => {
+export const success = async (ref: DocumentReference, response: Record<string, any>): Promise<WriteResult> => {
   return await ref.update({
-    config: config,
+    config: Config.json(),
     response:
     {
       ...{
@@ -30,6 +30,19 @@ export const success = async (ref: DocumentReference, response: Record<string, a
         timestamp: FieldValue.serverTimestamp(),
       },
       ...response,
+    },
+  });
+}
+
+export const failure = async (ref: DocumentReference, code: string, message: string): Promise<WriteResult> => {
+  return await ref.update({
+    config: Config.json(),
+    response:
+    {
+      status: "error",
+      code,
+      message,
+      timestamp: FieldValue.serverTimestamp(),
     },
   });
 }